requirements.txt
# Core dependencies
streamlit>=1.30.0
python-dotenv>=1.0.0

# LangChain and LangGraph
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20

# Model integrations
langchain-openai>=0.0.5
langchain-anthropic>=0.0.5

# Vector stores
langchain-astradb>=0.0.5
faiss-cpu>=1.7.4

# Search tools
langchain-tavily>=0.0.5

# Document processing
pypdf>=3.15.1
pillow>=10.0.0

# Optional: for development
pytest>=7.4.0
black>=23.7.0
